Rename files, edit testing.md
