+++ client/modules/App/App.test.jsx
@@ -12,7 +12,7 @@ import {
   within,
   prettyDOM
 } from '../../test-utils';
-import { initialTestState } from '../../redux_test_stores/test_store';
+import { initialTestState } from '../../testData/testReduxStore';
 
 jest.mock('../../i18n');
